skip remaining samples after reading only a subrange
The `read_*_samples()` functions did not leave the `io` read position
at the end of the `"data"` chunk if only a subrange that doesn't include
the last sample was read, which caused parsing of subsequent chunks
to fail. We therefore now pass on `chunk_size` to all the `read_*_sample()`
functions, such that they all can now always skip to the end of the chunk.