remove compile-time find_library() checks for presence of shared library
If this package gets precompiled and cached before the user has
installed a pre-requisite shared library, then installing said library
will not fix the problem if the check for the presence happened at
compile time. Likewise, if PackageCompiler.jl is used to compile an app
on another machine, then absence of a required shared library at
compile time there cannot be rectified later by a user who downloaded
a pre-compiled app containing this package. Dropping the compile-time
check for the presence of the required shared libraries fixes this.
The exception raised by ccall() if the required library is missing is
anyway more informative than the warning messages removed here.

https://julialang.github.io/PackageCompiler.jl/dev/devdocs/relocatable_part_3/